Relocate types
