+++ src/config/defaultConfig.ts
@@ -3,7 +3,7 @@
  * While editing this file, please keep in mind that these values are
  * completely visible to the user. Do not keep secrets here!
  */
-import { Endpoints } from "../enums";
+import { Endpoints } from "../types";
 
 export const defaultConfig = {
   service: "habits",
