+++ src/services/axios/index.ts
@@ -5,7 +5,7 @@ import {
   storeTokensOnReject,
 } from "./storeTokensInterceptor";
 import { appendTokensOnFulfill } from "./appendTokensInterceptor";
-import { Endpoints } from "../../enums";
+import { Endpoints } from "../../types";
 
 declare module "axios" {
   interface AxiosResponse {
@@ -18,7 +18,7 @@ declare module "axios" {
 }
 
 export const mainApi = axios.create({
-  baseURL: Endpoints[userConfig.get("endpoint")] || defaultConfig.endpoint,
+  baseURL: Endpoints[userConfig.get("endpointName")] || defaultConfig.endpoint,
   timeout: 20 * 1000,
 });
