+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./spinners";
 export * from "./commander";
 export * from "./secrets";
 export * from "./api";
+export * from "./config";
